Image_bot/requirements.txt
streamlit
pandas
openai
Pillow
python-dotenv